rtc: tps65910: Inherit wakeup state from parent device
TPS65910 is MFD device and RTC is one of its functions. The wakeup-source
property is applied to the parent MFD device and we need to apply it to
the RTC device as well, otherwise RTC can't be used for waking up system
from suspend by default, which is the expected behaviour. Inherit wakeup
state from the parent MFD driver.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>